requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
boto3==1.39.3
botocore==1.39.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
distro==1.9.0
Django==5.2.3
djangorestframework==3.16.0
dotenv==0.9.9
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
openai==1.91.0
packaging==25.0
playwright==1.53.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
PyJWT==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
